LayoutTests/http/tests/inspector/indexeddb/database-names-expected.txt

CONSOLE MESSAGE: line 108: InspectorTest.IndexedDB_callback1
CONSOLE MESSAGE: line 108: InspectorTest.IndexedDB_callback2
CONSOLE MESSAGE: line 108: InspectorTest.IndexedDB_callback3
CONSOLE MESSAGE: line 108: InspectorTest.IndexedDB_callback4
Tests that database names are correctly loaded and saved in IndexedDBModel.

Dumping database names:

Dumping database names:
    testDatabase1

Dumping database names:
    testDatabase1
    testDatabase2

Dumping database names:
    testDatabase1

Dumping database names:
